chore: split hooks.js up into separate files and begin some related tests
