+++ src/react/hooks/index.js
@@ -1,0 +1,3 @@
+export { default as useAppEvent } from './useAppEvent';
+
+export * from './paragon';
