+++ src/react/hooks/paragon/index.js
@@ -1,0 +1,2 @@
+export { default as useTrackColorSchemeChoice } from './useTrackColorSchemeChoice';
+export { default as useParagonTheme } from './useParagonTheme';
