+++ src/react/hooks/paragon/useTrackColorSchemeChoice.test.js
@@ -1,8 +1,13 @@
 import { renderHook } from '@testing-library/react-hooks';
-import { useTrackColorSchemeChoice } from './hooks';
-import { sendTrackEvent } from '../analytics';
 
-jest.mock('../analytics');
+import { sendTrackEvent } from '../../../analytics';
+
+import useTrackColorSchemeChoice from './useTrackColorSchemeChoice';
+
+jest.mock('../../../analytics', () => ({
+  ...jest.requireActual('../../../analytics'),
+  sendTrackEvent: jest.fn(),
+}));
 
 const mockAddEventListener = jest.fn();
 const mockRemoveEventListener = jest.fn();
@@ -16,7 +21,7 @@ Object.defineProperty(window, 'matchMedia', {
   })),
 });
 
-describe('useTrackColorSchemeChoice hook', () => {
+describe('useTrackColorSchemeChoice', () => {
   afterEach(() => {
     mockAddEventListener.mockClear();
     mockRemoveEventListener.mockClear();
